move example to correct folder
