+++ example_Basics/src/testApp.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "ofMain.h"
+#include "ofxMioFlowGLSL.h"
+
+
+class testApp : public ofBaseApp{  
+public:
+    void setup();
+    void update();
+    void draw();
+
+    ofVideoGrabber cam;
+    ofxMioFlowGLSL mioFlow;
+};
