--- example_Basics/testApp.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#include "ofMain.h"
-#include "ofxMioFlowGLSL.h"
-
-
-class testApp : public ofBaseApp{  
-    public:  
-        void setup();  
-        void update();  
-        void draw();  
-
-      
-        ofVideoGrabber cam;  
-		ofxMioFlowGLSL mioFlow;
-};
